# Label performance improvements and refactoring