--- a/tests/unit/text/test_caret.py
+++ b/tests/unit/text/test_caret.py
@@ -41,9 +41,10 @@ def mock_layout():
     _layout.foreground_decoration_group.attach_mock(program, 'program')
 
     # Allow the shader program to create a mock vertex list on demand
-    def _fake_vertex_list_method(count, mode, batch=None, group=None, colors=None):
+    def _fake_vertex_list_method(count, mode, batch=None, group=None, colors=None, visible=None):
         vertex_list = NonCallableMock(spec=IndexedVertexList)
         vertex_list.colors = ListSlicesAsTuple(colors[1])
+        vertex_list.visible = (1, 1)
         return vertex_list
 
     program.vertex_list = _fake_vertex_list_method